+++ configs/vite.config.ts
@@ -14,4 +14,7 @@ export default defineConfig({
     outDir: "dist",
     emptyOutDir: true,
   },
+  worker: {
+    format: "es",
+  },
 });